Restructurar componentes y estilos con folders
