+++ src/components/Preloader/Preloader.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import "../blocks/Preloader.css";
+//import "../blocks/Preloader.css";
 
 function Preloader() {
   return (
