+++ src/index.css
@@ -1,10 +1,10 @@
 @import url("./blocks/normalize.css");
 @import url("./blocks/fonts.css");
-@import url("./blocks/App.css");
-@import url("./blocks/Homepage.css");
-@import url("./blocks/Header.css");
-@import url("./blocks/Profile.css");
-@import url("./blocks/Elements.css");
-@import url("./blocks/Popup.css");
-@import url("./blocks/Preloader.css");
-@import url("./blocks/Footer.css");
+@import url("./components/App/App.css");
+@import url("./components/Home/Home.css");
+@import url("./components/Header/Header.css");
+@import url("./components/Navigation/Navigation.css");
+@import url("./components/Main/Main.css");
+@import url("./components/ImagePopup/ImagePopup.css");
+@import url("./components/Preloader/Preloader.css");
+@import url("./components/Footer/Footer.css");
